requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
elastic-transport==8.4.0
elasticsearch==8.4.1
filelock==3.8.0
idna==3.3
lxml==4.9.1
numpy==1.23.3
pandas==1.4.4
pymongo==4.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
urllib3==1.26.12
